change key type of mappers to int64_t

+++ examples/bi_gram_mapper.cc
@@ -5,7 +5,7 @@
 #include "bi_gram_mapper.h"
 
 void BiGramMapper::Map(google::dense_hash_map<std::string, std::vector<int>> &kvs,
-                       const long long int &key,
+                       const int64_t &key,
                        const std::string &value) {
   size_t cur = 0, p1, p2 = 0;
   p1 = value.find_first_of(' ', cur);
