+++ examples/group_by_mapper.cc
@@ -5,7 +5,7 @@
 #include "group_by_mapper.h"
 
 void GroupByMapper::Map(google::dense_hash_map<int, std::vector<std::string>> &kvs,
-                        const long long int &key,
+                        const int64_t &key,
                         const std::string &value) {
   size_t cur = 0, pos;
   while ((pos = value.find_first_of(' ', cur)) != value.npos) {
