+++ examples/group_by_mapper.h
@@ -7,17 +7,17 @@
 
 #include <mapper.h>
 
-class GroupByMapper: public Mapper<int, std::string, long long int, std::string> {
+class GroupByMapper: public Mapper<int, std::string, int64_t, std::string> {
 
  public:
   virtual void Map(
       google::dense_hash_map<int, std::vector<std::string>> &kvs,
-      const long long int &key,
+      const int64_t &key,
       const std::string &value) override;
 };
 
-extern "C" std::unique_ptr<Mapper<int, std::string, long long int, std::string>> Create() {
-  return std::unique_ptr<Mapper<int, std::string, long long int, std::string>>(new GroupByMapper);
+extern "C" std::unique_ptr<Mapper<int, std::string, int64_t, std::string>> Create() {
+  return std::unique_ptr<Mapper<int, std::string, int64_t, std::string>>(new GroupByMapper);
 }
